GO-5593 move autoInstallSpaceChatWidget func
